+++ src/main/gen/com/github/nthykier/debpkg/deb822/deplang/psi/impl/DepLangArchRestrictionPartImpl.java
@@ -22,6 +22,7 @@ public class DepLangArchRestrictionPartImpl extends ASTWrapperPsiElement impleme
     visitor.visitArchRestrictionPart(this);
   }
 
+  @Override
   public void accept(@NotNull PsiElementVisitor visitor) {
     if (visitor instanceof DepLangVisitor) accept((DepLangVisitor)visitor);
     else super.accept(visitor);