+++ src/main/gen/com/github/nthykier/debpkg/deb822/deplang/psi/impl/DepLangSubstvarImpl.java
@@ -23,6 +23,7 @@ public class DepLangSubstvarImpl extends ASTWrapperPsiElement implements DepLang
     visitor.visitSubstvar(this);
   }
 
+  @Override
   public void accept(@NotNull PsiElementVisitor visitor) {
     if (visitor instanceof DepLangVisitor) accept((DepLangVisitor)visitor);
     else super.accept(visitor);