+++ src/main/gen/com/github/nthykier/debpkg/deb822/deplang/psi/impl/DepLangVersionOperatorImpl.java
@@ -22,6 +22,7 @@ public class DepLangVersionOperatorImpl extends ASTWrapperPsiElement implements 
     visitor.visitVersionOperator(this);
   }
 
+  @Override
   public void accept(@NotNull PsiElementVisitor visitor) {
     if (visitor instanceof DepLangVisitor) accept((DepLangVisitor)visitor);
     else super.accept(visitor);