+++ src/main/java/com/github/nthykier/debpkg/Deb822Bundle.java
@@ -1,11 +1,11 @@
 package com.github.nthykier.debpkg;
 
-import com.intellij.AbstractBundle;
+import com.intellij.DynamicBundle;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.PropertyKey;
 
-public class Deb822Bundle extends AbstractBundle {
+public class Deb822Bundle extends DynamicBundle {
 
     @NonNls
     public static final String BUNDLE_NAME = "messages.DebpkgBundle";