+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - When dependency profiles are present in a dependency field where it is not permitted, flag the
   profile(s) and not the entire dependency as problematic.
 - Bump build time dependencies (org.jetbrains.intellij, org.jetbrains.changelog)
+- Internal/Chore: Bump gradle version to 7.3.
 
 ## [0.0.12]
 ### Changed